+++ src/main/java/com/halggeol/backend/user/mapper/UserMapper.java
@@ -1,6 +1,6 @@
 package com.halggeol.backend.user.mapper;
 
-import com.halggeol.backend.security.account.domain.User;
+import com.halggeol.backend.security.domain.User;
 
 public interface UserMapper {
     User findByEmail(String email);
