+++ src/main/java/com/halggeol/backend/user/service/UserServiceImpl.java
@@ -1,6 +1,6 @@
 package com.halggeol.backend.user.service;
 
-import com.halggeol.backend.security.account.domain.User;
+import com.halggeol.backend.security.domain.User;
 import com.halggeol.backend.security.mail.domain.MailType;
 import com.halggeol.backend.security.mail.dto.MailDTO;
 import com.halggeol.backend.security.mail.service.MailService;
